Use libsndfile and floats

+++ src/main/gui/Waveform.h
@@ -28,11 +28,11 @@ private:
     bool shouldShowSamplePoints(const double samplesPerPixel, const uint8_t hardwarePixelsPerAppPixel);
 
     std::vector<std::unique_ptr<SamplePoint>> computeSamplePoints(int width, int height,
-                                 const std::vector<int16_t>& samples, size_t offset,
+                                 const std::vector<float>& samples, size_t offset,
                                  float samplesPerPixel, float verticalZoom, const uint8_t hardwarePixelsPerAppPixel);
     
     void renderSmoothWaveform(SDL_Renderer* renderer, int width, int height,
-                                     const std::vector<int16_t>& samples, size_t offset,
+                                     const std::vector<float>& samples, size_t offset,
                                      float samplesPerPixel, float verticalZoom, const uint8_t hardwarePixelsPerAppPixel);
 
     void handleScroll(const int32_t mouseX,
